requirements.txt
nipype==1.3.1
pybids==0.9.5
